@@ -3,7 +3,7 @@ package cl.ceisufro.fluttermbtilesextractor
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 
 object Utils {
     fun hasPermissions(context: Context): Boolean {
